requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==40.0.2
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
jwt==1.3.1
pycparser==2.21
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
